pyproject.toml
[tool.poetry]
name = "dynatable"
version = "0.1.0"
description = ""
authors = ["blooser <blooser@protonmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.3"
djangorestframework = "^3.15.1"
pytest-django = "^4.8.0"
sqlalchemy = "^2.0.29"
postgres = "^4.0"
shortuuid = "^1.0.13"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
